requirements.txt
requests
GitPython
openai
click
pyperclip